Add JavaDoc to project files.
