+++ text-ui-test/EXPECTED.TXT
@@ -12,7 +12,7 @@ Hello from
      \/__/   \/__/
 
     ---------------------------------------------------------------------******************CHATTIME
-     Hey! I'm your friend, Chattime!  (•◡•) /
+     Hey! I'm your friend, chattime.Chattime!  (•◡•) /
      How can I help you *^*
     ---------------------------------------------------------------------******************CHATTIME
     ---------------------------------------------------------------------******************CHATTIME
